change client id related type from int to uintptr_t

+++ src/GameClient.cpp
@@ -2,8 +2,8 @@
 #include <iostream>
 
 
-GameClient::GameClient(int clientID)
-    : m_clientID(clientID) {}
+GameClient::GameClient(uintptr_t clientID)
+        : m_clientID(clientID) {}
 
 
 Message GameClient::prepareMessageToServer(const Message& message) const{
@@ -13,8 +13,8 @@ Message GameClient::prepareMessageToServer(const Message& message) const{
 
 void GameClient::onMessageFromServer(Message& message)
 {
-    std::cout << "[GameClient " << m_clientID 
-    << "] Received message from server\n";
+    std::cout << "[GameClient " << m_clientID
+              << "] Received message from server\n";
     // TODO
 }
